apps/backend/pyproject.toml
[project]
name = "persona-mirror-backend"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "passlib>=1.7.4,<2.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pyjwt>=2.9.0,<3.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]
